+++ packages/admin/src/components/AdminContainer.tsx
@@ -6,8 +6,7 @@ import { Redirect, Route, Switch } from "react-router";
 import styled from "styled-components/macro";
 import Modules, { ReduxStoreState } from "modules/root";
 import { Link } from "react-router-dom";
-import { COLORS, MOBILE } from "tools/constants";
-import { HEADER_HEIGHT } from "tools/dimensions";
+import { COLORS, HEADER_HEIGHT, MOBILE } from "tools/constants";
 import AdminSummary from "./AdminSummary";
 import Index from "./Index";
 import Swipy from "swipyjs";
