+++ packages/admin/src/components/AdminNavigationMenu.tsx
@@ -2,9 +2,8 @@ import React from "react";
 import { connect } from "react-redux";
 import styled from "styled-components/macro";
 import Modules, { ReduxStoreState } from "modules/root";
-import { COLORS, MOBILE } from "tools/constants";
-import { HEADER_HEIGHT } from "tools/dimensions";
-import { composeWithProps } from "tools/utils";
+import { COLORS, HEADER_HEIGHT, MOBILE } from "tools/constants";
+import { composeWithProps } from "tools/admin-utils";
 import { Button } from "@blueprintjs/core";
 import { NavLink, RouteComponentProps } from "react-router-dom";
 import cx from "classnames";
