+++ packages/admin/src/tools/admin-utils.ts
@@ -1,4 +1,162 @@
+import { compose } from "redux";
+import { CourseList, CourseSkeleton } from "@pairwise/common";
+import { ParsedQuery } from "query-string";
 import { AdminUserView } from "../modules/users/store";
+
+/**
+ * Artificially wait the provided amount of time.
+ */
+export const wait = async (time: number = 1000) => {
+  await new Promise((_: any) => setTimeout(_, time));
+};
+
+/**
+ * A compose function which passes only props information for better
+ * type-checking.
+ */
+export const composeWithProps = <T extends {}>(
+  ...fns: any
+): ((component: any) => (props: T) => any) => {
+  return compose(...fns);
+};
+
+type OS = "Mac" | "iOS" | "Windows" | "Android" | "Linux";
+
+/**
+ * Determine client OS, code from this StackOverflow post:
+ * https://stackoverflow.com/questions/38241480/detect-macos-ios-windows-android-and-linux-os-with-js
+ */
+export const getClientOS = (): Nullable<OS> => {
+  const userAgent = window.navigator.userAgent;
+  const platform = window.navigator.platform;
+  const macPlatforms = new Set(["Macintosh", "MacIntel", "MacPPC", "Mac68K"]);
+  const windowsPlatforms = new Set(["Win32", "Win64", "Windows", "WinCE"]);
+  const iosPlatforms = new Set(["iPhone", "iPad", "iPod"]);
+  let os: Nullable<OS> = null;
+
+  if (macPlatforms.has(platform)) {
+    os = "Mac";
+  } else if (iosPlatforms.has(platform)) {
+    os = "iOS";
+  } else if (windowsPlatforms.has(platform)) {
+    os = "Windows";
+  } else if (/Android/.test(userAgent)) {
+    os = "Android";
+  } else if (!os && /Linux/.test(platform)) {
+    os = "Linux";
+  }
+
+  return os;
+};
+
+/**
+ * Find a course by id in the course list.
+ */
+export const findCourseById = (courseId: string, courses: CourseList) => {
+  const course = courses.find(c => c.id === courseId);
+  return course;
+};
+
+/**
+ * Map the course skeletons when fetched in DEV mode. This just transforms
+ * the skeletons so all the content is available in development.
+ */
+export const mapCourseSkeletonInDev = (courseSkeleton: CourseSkeleton) => {
+  return {
+    ...courseSkeleton,
+    modules: courseSkeleton.modules.map(m => {
+      return {
+        ...m,
+        free: true,
+        userCanAccess: true,
+        challenges: m.challenges.map(c => {
+          return {
+            ...c,
+            userCanAccess: true,
+          };
+        }),
+      };
+    }),
+  };
+};
+
+// Definition of specific types to identify different states for the
+// initial app load:
+export enum APP_INITIALIZATION_TYPE {
+  DEFAULT = "DEFAULT",
+
+  // Special types:
+  SIGN_IN = "SIGN_IN",
+  EMAIL_UPDATED = "EMAIL_UPDATED",
+  ACCOUNT_CREATED = "ACCOUNT_CREATED",
+  PAYMENT_SUCCESS = "PAYMENT_SUCCESS",
+  PAYMENT_CANCELLED = "PAYMENT_CANCELLED",
+  PURCHASE_COURSE_FLOW = "PURCHASE_COURSE_FLOW",
+  AUTHENTICATION_FAILURE = "AUTHENTICATION_FAILURE",
+}
+
+// Check a list of param keys exists in the parsed query params from a url
+const checkParamsExist = (params: ParsedQuery<string>, keys: string[]) => {
+  for (const key of keys) {
+    if (params[key] === undefined) {
+      return false;
+    }
+  }
+
+  return true;
+};
+
+// Parse the initial url and params which may have loaded the app and
+// return the APP_INITIALIZATION_TYPE enum which specifically
+// categorizes the app initialization type. This is then used to trigger
+// other events in the app based on the initialization state.
+export const parseInitialUrlToInitializationType = (
+  path: string,
+  params: ParsedQuery<string>,
+): APP_INITIALIZATION_TYPE => {
+  // A user authenticated:
+  if (checkParamsExist(params, ["accessToken", "accountCreated"])) {
+    if (params.accountCreated === "true") {
+      return APP_INITIALIZATION_TYPE.ACCOUNT_CREATED;
+    } else {
+      return APP_INITIALIZATION_TYPE.SIGN_IN;
+    }
+  }
+
+  // A user updated their email successfully:
+  if (path === "/account" && checkParamsExist(params, ["emailUpdated"])) {
+    if (params.emailUpdated === "true") {
+      return APP_INITIALIZATION_TYPE.EMAIL_UPDATED;
+    }
+  }
+
+  // There was some error during authentication.
+  if (
+    path === "/authentication-failure" &&
+    checkParamsExist(params, ["strategy"])
+  ) {
+    return APP_INITIALIZATION_TYPE.AUTHENTICATION_FAILURE;
+  }
+
+  // A deep link into the payment flow, which also expects a courseId
+  // param, e.g. https://app.pairwise.tech/purchase?courseId="fpvPtfu7s"
+  if (path === "/purchase") {
+    return APP_INITIALIZATION_TYPE.PURCHASE_COURSE_FLOW;
+  }
+
+  // A user returned from the checkout flow after cancelling:
+  if (path === "/payment-cancelled") {
+    return APP_INITIALIZATION_TYPE.PAYMENT_CANCELLED;
+  }
+
+  // A user returned from the checkout flow after payment success:
+  if (path === "/payment-success" && checkParamsExist(params, ["courseId"])) {
+    return APP_INITIALIZATION_TYPE.PAYMENT_SUCCESS;
+  }
+
+  // Default category:
+  return APP_INITIALIZATION_TYPE.DEFAULT;
+};
 
 /**
  * Parse the course progress.
